requirements.txt
Jinja2>=3.1.0
PyYAML>=6.0
yt-dlp>=2025.1.0
youtube-transcript-api>=0.6.0
